refactor(tracers): clean up and reorganize block tracer code for improved readability and maintainability
