+++ core/events/trace.go
@@ -1,7 +1,9 @@
 package events
 
-import "github.com/harmony-one/harmony/hmy/tracers"
+import (
+	"github.com/harmony-one/harmony/hmy/tracers/native"
+)
 
 type TraceEvent struct {
-	Tracer *tracers.ParityBlockTracer
+	Tracer *native.ParityBlockTracer
 }
