+++ hmy/tracers/rosetta_block_tracer.go
@@ -31,64 +31,3 @@ type RosettaLogItem struct {
 	To        *vm.RosettaLogAddressItem
 	Value     *big.Int
 }
-
-type RosettaBlockTracer struct {
-	*ParityBlockTracer
-
-	logs []*RosettaLogItem
-}
-
-func (rbt *RosettaBlockTracer) formatAction(depth []int, parentErr error, ac *action) *RosettaLogItem {
-	val := ac.value
-	if val == nil {
-		val = big.NewInt(0)
-	}
-
-	return &RosettaLogItem{
-		IsSuccess: ac.err == nil,
-		Reverted:  !(parentErr == nil && ac.err == nil),
-		OP:        ac.op,
-		Depth:     depth,
-		From:      &vm.RosettaLogAddressItem{Account: &ac.from},
-		To:        &vm.RosettaLogAddressItem{Account: &ac.to},
-		Value:     val,
-	}
-}
-
-func (rbt *RosettaBlockTracer) AddRosettaLog(op vm.OpCode, from, to *vm.RosettaLogAddressItem, val *big.Int) {
-	rbt.logs = append(rbt.logs, &RosettaLogItem{
-		IsSuccess: true,
-		Reverted:  false,
-		OP:        op,
-		Depth:     make([]int, 0),
-		From:      from,
-		To:        to,
-		Value:     val,
-	})
-}
-
-func (rbt *RosettaBlockTracer) GetResult() ([]*RosettaLogItem, error) {
-	root := &rbt.cur.action
-
-	var results = make([]*RosettaLogItem, 0)
-	var err error
-	var finalize func(ac *action, parentErr error, traceAddress []int)
-	finalize = func(ac *action, parentErr error, traceAddress []int) {
-		results = append(results, rbt.formatAction(traceAddress, parentErr, ac))
-		nextErr := parentErr
-		if ac.err != nil {
-			nextErr = ac.err
-		}
-
-		for i, subAc := range ac.subCalls {
-			finalize(subAc, nextErr, append(traceAddress[:], i))
-		}
-	}
-
-	traceAddress := make([]int, 0)
-	for i, subAc := range root.subCalls {
-		finalize(subAc, root.err, append(traceAddress[:], i))
-	}
-
-	return append(results, rbt.logs...), err
-}
